mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "space-apps-mcp"
version = "0.1.0"
description = "MCP server for NASA Space Apps Chicago data exploration tools"
authors = [
    { name = "NASA Space Apps Chicago Team" }
]
dependencies = [
    "fastmcp>=0.1.0",
    "elasticsearch>=8.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.1.1",
    "isort>=5.13.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/space_apps_mcp"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"

[project.urls]
Homepage = "https://github.com/nasa-space-apps-chicago"
Repository = "https://github.com/nasa-space-apps-chicago/mcp-server"
Issues = "https://github.com/nasa-space-apps-chicago/mcp-server/issues"

[project.scripts]
elastic-appears-mcp = "elastic_appears_mcp.cli:main" 